refactured and created new folders
